+++ backend/canisters/group/impl/src/lifecycle/post_upgrade.rs
@@ -1,6 +1,7 @@
 use crate::lifecycle::{init_env, init_state, UPGRADE_BUFFER_SIZE};
 use crate::memory::get_upgrades_memory;
-use crate::Data;
+use crate::model::upgrade_instruction_counts::InstructionCountFunctionId;
+use crate::{read_state, Data};
 use canister_logger::LogEntry;
 use canister_tracing_macros::trace;
 use group_canister::post_upgrade::Args;
@@ -29,4 +30,11 @@ fn post_upgrade(args: Args) {
     init_state(env, data, args.wasm_version);
 
     info!(version = %args.wasm_version, "Post-upgrade complete");
+
+    read_state(|state| {
+        let now = state.env.now();
+        state
+            .data
+            .record_instructions_count(InstructionCountFunctionId::PostUpgrade, now)
+    });
 }